requirements.txt
fastapi==0.115.12
httpx==0.28.1
PyJWT==2.10.1
pytest==8.3.5
psycopg[binary]==3.2.9
redis==6.0.0
uvicorn==0.34.2
bcrypt==5.0.0
